[refactor] リポジトリのデータモデルのDiffUtilをモデルのktファイルまで一時退避
