+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/domain/model/RepositoryModel.kt
@@ -1,6 +1,7 @@
 package jp.co.yumemi.android.code_check.domain.model
 
 import android.os.Parcelable
+import androidx.recyclerview.widget.DiffUtil
 import kotlinx.parcelize.Parcelize
 
 
@@ -17,3 +18,21 @@ data class RepositoryDataModel(
     val forksCount: Long,
     val openIssuesCount: Long,
 ) : Parcelable
+
+
+val repositoryDiffUtil = object : DiffUtil.ItemCallback<RepositoryDataModel>() {
+    override fun areItemsTheSame(
+        oldItem: RepositoryDataModel,
+        newItem: RepositoryDataModel
+    ): Boolean {
+        return oldItem.name == newItem.name
+    }
+
+    override fun areContentsTheSame(
+        oldItem: RepositoryDataModel,
+        newItem: RepositoryDataModel
+    ): Boolean {
+        return oldItem == newItem
+    }
+
+}
