+++ app/src/main/kotlin/jp/co/yumemi/android/code_check/ui/search/SearchScreenFragment.kt
@@ -8,7 +8,6 @@ import android.view.View
 import android.view.inputmethod.EditorInfo
 import androidx.fragment.app.Fragment
 import androidx.navigation.fragment.findNavController
-import androidx.recyclerview.widget.DiffUtil
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import jp.co.yumemi.android.code_check.R
@@ -70,21 +69,4 @@ class SearchScreenFragment : Fragment(R.layout.fragment_search_screen) {
             SearchScreenFragmentDirections.actionRepositoriesFragmentToRepositoryFragment(item = item)
         findNavController().navigate(action)
     }
-}
-
-val diff_util = object : DiffUtil.ItemCallback<RepositoryDataModel>() {
-    override fun areItemsTheSame(
-        oldItem: RepositoryDataModel,
-        newItem: RepositoryDataModel
-    ): Boolean {
-        return oldItem.name == newItem.name
-    }
-
-    override fun areContentsTheSame(
-        oldItem: RepositoryDataModel,
-        newItem: RepositoryDataModel
-    ): Boolean {
-        return oldItem == newItem
-    }
-
 }
